ANDROID: arm64: Fix off-by-one vdso trampoline return value
The VDSO_compat_* symbols added by vdso.lds.S for sigreturn32 are not
being tagged in the symbol table as FUNC symbols by ld.lld.  As
a result, I see a SIGBUS error when running the LTP test
syscalls.rt_sigaction01.

This was fixed in https://reviews.llvm.org/D86263. Until we can update
to the latest version of llvm, let's drop the bottom bit for these
symbols in the VDSO_SYMBOL() macro so that the `vdso_trampoline + thumb`
functions properly.

Test: LTP test syscalls.rt_sigaction01
Bug: 172307050
Fixes: 71ca95b7e949 ("FROMLIST: arm64: vdso32: Allow ld.lld to properly link the VDSO")
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I24cf58ae62616a80c82653a7fc7dea93b33a3d0d
Signed-off-by: atndko <z1281552865@gmail.com>
Signed-off-by: Zlatan Radovanovic <zlatan.radovanovic@fet.ba>